LayoutTests/http/tests/security/contentSecurityPolicy/worker-multiple-csp-headers-expected.txt

CONSOLE MESSAGE: Refused to connect to 'http://127.0.0.1:8000/xmlhttprequest/resources/get.txt' because it violates the following Content Security Policy directive: "connect-src 'none'".


PASS Worker can have multiple CSP headers 
